Group classes into packages
Group classes from A-MoreOOP into their respective packages and rearrange the order of the imported classes alphabetically
