--- src/main/java/commands/ListCommand.java
@@ -1,29 +1,0 @@
-package commands;
-
-import exceptions.DukeException;
-import storage.Storage;
-import tasks.TaskList;
-import ui.Ui;
-
-/**
- * Adds a person to the address book.
- */
-public class ListCommand extends Command {
-
-    public static final String COMMAND_WORD = "list";
-
-    public static final String MESSAGE_SUCCESS = "Listed below are the tasks I remember: ";
-
-    public ListCommand() {
-    }
-
-    @Override
-    public boolean isByeCommand() {
-        return false;
-    }
-
-    @Override
-    public void execute(TaskList tasks, Ui ui, Storage storage) throws DukeException {
-        ui.showMessage(MESSAGE_SUCCESS + System.lineSeparator() + tasks.showList());
-    }
-}
