+++ src/main/java/duke/commands/ListCommand.java
@@ -1,0 +1,29 @@
+package duke.commands;
+
+import duke.exceptions.DukeException;
+import duke.storage.Storage;
+import duke.tasks.TaskList;
+import duke.ui.Ui;
+
+/**
+ * Adds a person to the address book.
+ */
+public class ListCommand extends Command {
+
+    public static final String COMMAND_WORD = "list";
+
+    public static final String MESSAGE_SUCCESS = "Listed below are the duke.tasks I remember: ";
+
+    public ListCommand() {
+    }
+
+    @Override
+    public boolean isByeCommand() {
+        return false;
+    }
+
+    @Override
+    public void execute(TaskList tasks, Ui ui, Storage storage) throws DukeException {
+        ui.showMessage(MESSAGE_SUCCESS + System.lineSeparator() + tasks.showList());
+    }
+}
