+++ src/main/java/duke/exceptions/DukeEmptyDescriptionException.java
@@ -1,0 +1,9 @@
+package duke.exceptions;
+
+public class DukeEmptyDescriptionException extends DukeException {
+
+    public DukeEmptyDescriptionException() {
+        super("Commands todo, deadline and event should have a description");
+    }
+
+}
