+++ src/main/java/duke/exceptions/DukeInvalidCommandException.java
@@ -1,0 +1,8 @@
+package duke.exceptions;
+
+public class DukeInvalidCommandException extends DukeException {
+
+    public DukeInvalidCommandException() {
+        super("You have entered an invalid command");
+    }
+}
