+++ src/main/java/duke/exceptions/DukeInvalidDateException.java
@@ -1,0 +1,8 @@
+package duke.exceptions;
+
+public class DukeInvalidDateException extends DukeException {
+
+    public DukeInvalidDateException() {
+        super("An invalid date has been entered. duke.tasks.Date should be formatted as YYYY-MM-DD");
+    }
+}
