+++ src/main/java/duke/exceptions/DukeInvalidFormatException.java
@@ -1,0 +1,9 @@
+package duke.exceptions;
+
+public class DukeInvalidFormatException extends DukeException {
+
+    public DukeInvalidFormatException(String type, String missingExpr) {
+        super(type + " should have a " + missingExpr + " command and a specified date in its description");
+    }
+
+}
