+++ src/main/java/duke/exceptions/DukeInvalidNumberFormatException.java
@@ -1,0 +1,8 @@
+package duke.exceptions;
+
+public class DukeInvalidNumberFormatException extends DukeException {
+
+    public DukeInvalidNumberFormatException(String command) {
+        super("Invalid number format detected. Enter something like \"" + command + " 1 \"");
+    }
+}
