+++ src/main/java/duke/exceptions/DukeUnspecifiedNumberException.java
@@ -1,0 +1,8 @@
+package duke.exceptions;
+
+public class DukeUnspecifiedNumberException extends DukeException {
+
+    public DukeUnspecifiedNumberException(String command) {
+        super(command + " command should be specified with a task number");
+    }
+}
