+++ src/main/java/duke/tasks/Task.java
@@ -1,0 +1,36 @@
+package duke.tasks;
+
+public class Task {
+    protected String description;
+    protected boolean isDone;
+
+    public Task(String description) {
+        this.description = description;
+        this.isDone = false;
+    }
+
+    public String getStatusIcon() {
+        return (isDone ? "X" : " "); // mark done task with X
+    }
+
+    @Override
+    public String toString() {
+        return "[" + getStatusIcon() + "] " + description;
+    }
+
+    public void markAsDone() {
+        this.isDone = true;
+    }
+
+    public void markAsNotDone() {
+        this.isDone = false;
+    }
+
+    public String toStringForFile() {
+        if (this.isDone) {
+            return "M|";
+        } else {
+            return "N|";
+        }
+    }
+}
