--- src/main/java/exceptions/DukeEmptyDescriptionException.java
@@ -1,9 +1,0 @@
-package exceptions;
-
-public class DukeEmptyDescriptionException extends DukeException {
-
-    public DukeEmptyDescriptionException() {
-        super("Commands todo, deadline and event should have a description");
-    }
-
-}
