--- src/main/java/exceptions/DukeInvalidCommandException.java
@@ -1,8 +1,0 @@
-package exceptions;
-
-public class DukeInvalidCommandException extends DukeException {
-
-    public DukeInvalidCommandException() {
-        super("You have entered an invalid command");
-    }
-}
