--- src/main/java/exceptions/DukeInvalidDateException.java
@@ -1,8 +1,0 @@
-package exceptions;
-
-public class DukeInvalidDateException extends DukeException {
-
-    public DukeInvalidDateException() {
-        super("An invalid date has been entered. tasks.Date should be formatted as YYYY-MM-DD");
-    }
-}
