--- src/main/java/exceptions/DukeInvalidFormatException.java
@@ -1,9 +1,0 @@
-package exceptions;
-
-public class DukeInvalidFormatException extends DukeException {
-
-    public DukeInvalidFormatException(String type, String missingExpr) {
-        super(type + " should have a " + missingExpr + " command and a specified date in its description");
-    }
-
-}
