--- src/main/java/exceptions/DukeInvalidNumberFormatException.java
@@ -1,8 +1,0 @@
-package exceptions;
-
-public class DukeInvalidNumberFormatException extends DukeException {
-
-    public DukeInvalidNumberFormatException(String command) {
-        super("Invalid number format detected. Enter something like \"" + command + " 1 \"");
-    }
-}
