--- src/main/java/exceptions/DukeUnspecifiedNumberException.java
@@ -1,8 +1,0 @@
-package exceptions;
-
-public class DukeUnspecifiedNumberException extends DukeException {
-
-    public DukeUnspecifiedNumberException(String command) {
-        super(command + " command should be specified with a task number");
-    }
-}
