--- src/main/java/tasks/Todo.java
@@ -1,19 +1,0 @@
-package tasks;
-
-public class Todo extends Task {
-
-    public Todo(String description) {
-        super(description);
-    }
-
-    @Override
-    public String toString() {
-        return "[T]" + super.toString();
-    }
-
-    @Override
-    public String toStringForFile() {
-        return super.toStringForFile() + "todo " + this.description;
-    }
-
-}
